backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synapse-backend"
version = "0.1.0"
description = "Ambient Relationship Intelligence OS Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "neo4j>=5.20.0",
    "redis>=5.0.0",
    "kafka-python>=2.0.0",
    "celery>=5.4.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "google-generativeai>=0.8.0",
    "web3>=6.20.0",
    "pyjwt>=2.8.0",
    "cryptography>=43.0.0",
    "sentry-sdk>=2.14.0",
    "tavily-python>=0.5.0",
    "pyttsx3>=2.90",
    "pytest-asyncio>=0.23.0"
]
